requirements.txt
Django>=4.2,<4.2.11
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9,<2.9.9
drf-spectacular>=0.26.1,<0.27.0
djangorestframework-simplejwt>=5.0.0,<5.3.0